Support decoding HCL via an Unmarshaler or TextUnmarshaler
Unmarshaler is a basic interface that any type can implement. It allows
the type being unmarshaled over to control how it's unmarshaled --
e.g., bit flags could be unmarshaled from a list and so on.

encoding.TextUnmarshaler supports a smaller range of literal token
types, namely only strings, heredocs, floats, numbers, and bools.
Unlike decoding an ast.Node into a string using DecodeObject, this will
use the text of floats and bools (whereas for DecodeObject, it's an
error to decode a float or bool into a string).